% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brunnermunzel.perutation.test.R
\name{brunnermunzel.permutation.test}
\alias{brunnermunzel.permutation.test}
\alias{brunnermunzel.permutation.test.default}
\alias{brunnermunzel.permutation.test.formula}
\title{permuted Brunner-Munzel test}
\usage{
brunnermunzel.permutation.test(x, ...)

\method{brunnermunzel.permutation.test}{default}(x, y,
  alternative = c("two.sided", "greater", "less"), ...)

\method{brunnermunzel.permutation.test}{formula}(formula, data, subset,
  na.action, ...)
}
\arguments{
\item{x}{the numeric vector of data values from the sample 1.}

\item{\dots}{further arguments to be passed to or from methods
(This argument is for only formula).}

\item{y}{the numeric vector of data values from the sample 2.}

\item{alternative}{a character string specifying the alternative
hypothesis, must be one of \code{"two.sided"} (default), \code{"greater"} or
\code{"less"}. User can specify just the initial letter.}

\item{formula}{a formula of the form \code{lhs ~ rhs} where \code{lhs}
is a numeric variable giving the data values and \code{rhs} a factor
with two levels giving the corresponding groups.}

\item{data}{an optional matrix or data frame (or similar: see
\code{\link{model.frame}}) containing the variables in the
formula \code{formula}.  By default the variables are taken from
\code{environment(formula)}.}

\item{subset}{an optional vector specifying a subset of observations
to be used.}

\item{na.action}{a function which indicates what should happen when
the data contain \code{NA}s.  Defaults to
\code{getOption("na.action")}.}
}
\value{
A list containing the following components:
 \item{method}{the characters ``permuted Brunner-Munzel Test''}
 \item{p.value}{the \eqn{p}-value of the test.}
 \item{data.name}{a character string giving the name of the data.}
}
\description{
This function performs the permuted Brunner-Munzel test.
}
\note{
FORTRAN subroutine `combination` in combination.f is derived from
the program by shikino (\url{http://slpr.sakura.ne.jp/qp/combination})
(CC-BY-4.0).
Thanks to shikono for your useful subroutine.
}
\examples{
Y <- c(1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 2, 4, 1, 1)
N <- c(3, 3, 4, 3, 1, 2, 3, 1, 1, 5, 4)

brunnermunzel.permutation.test(Y, N)

##       permuted Brunner-Munzel Test
##
## data:  Y and N
## p-value = 0.008038


## Formula interface.
dat <- data.frame(
    value = c(Y, N),
    group = factor(rep(c("Y", "N"), c(length(Y), length(N))),
                   levels = c("Y", "N"))
)

brunnermunzel.permutation.test(value ~ group, data = dat)

##       permuted Brunner-Munzel Test
##
## data:  value by group
## p-value = 0.008038

}
\references{
Karin Neubert and Edgar Brunner,
``A studentized permutation test for the non-parametric
Behrens-Fisher problem'',
Computational Statistics and Data Analysis, Vol. 51, pp. 5192-5204 (2007).
}
\seealso{
This function is made in reference to following cites (in Japanese):
Prof. Haruhiko Okumura
(\url{https://oku.edu.mie-u.ac.jp/~okumura/stat/brunner-munzel.html})
and Anonymous
(\url{https://blog.goo.ne.jp/r-de-r/e/83dc811baf41ecfe469fa794a4c51b84})
}
